tests: Move existing tests to tests/legacy/
