+++ tests/legacy/test_channel_creation.py
@@ -20,7 +20,7 @@ from nidaqmx.constants import (
     StrainGageBridgeType,
     BridgeConfiguration,
 )
-from .helpers import generate_random_seed
+from ..helpers import generate_random_seed
 
 
 class TestAnalogCreateChannels(object):
