+++ tests/legacy/test_container_operations.py
@@ -4,7 +4,7 @@ import random
 import pytest
 
 import nidaqmx
-from .helpers import generate_random_seed
+from ..helpers import generate_random_seed
 from nidaqmx.utils import flatten_channel_string
