+++ tests/legacy/test_enum_bitfields.py
