+++ tests/legacy/test_events.py
@@ -6,7 +6,7 @@ import pytest
 
 import nidaqmx
 from nidaqmx.constants import AcquisitionType
-from .helpers import generate_random_seed
+from ..helpers import generate_random_seed
 
 
 class TestEvents(object):
