+++ tests/legacy/test_invalid_reads.py
@@ -7,7 +7,7 @@
 # from nidaqmx.errors import DaqError
 # from nidaqmx.utils import flatten_channel_string
 # from .fixtures import any_x_series_device
-# from .helpers import generate_random_seed
+# from ..helpers import generate_random_seed
 # from .test_read_write import TestDAQmxIOBase
