+++ tests/legacy/test_power_up_states.py
