+++ tests/legacy/test_read_exceptions.py
