+++ tests/legacy/test_resource_warnings.py
