+++ tests/legacy/test_stream_analog_readers_writers.py
@@ -9,7 +9,7 @@ import nidaqmx
 from nidaqmx.constants import Edge
 from nidaqmx.stream_readers import AnalogSingleChannelReader, AnalogMultiChannelReader
 from nidaqmx.stream_writers import AnalogSingleChannelWriter, AnalogMultiChannelWriter
-from .helpers import generate_random_seed
+from ..helpers import generate_random_seed
 from .test_read_write import TestDAQmxIOBase
 from nidaqmx.utils import flatten_channel_string
