+++ tests/legacy/test_stream_digital_readers_writers.py
@@ -9,7 +9,7 @@ import nidaqmx
 from nidaqmx.constants import LineGrouping
 from nidaqmx.stream_readers import DigitalSingleChannelReader, DigitalMultiChannelReader
 from nidaqmx.stream_writers import DigitalSingleChannelWriter, DigitalMultiChannelWriter
-from .helpers import generate_random_seed
+from ..helpers import generate_random_seed
 from .test_read_write import TestDAQmxIOBase
 from nidaqmx.utils import flatten_channel_string
