+++ tests/legacy/test_write_exceptions.py
